A big commit of the project after the AA2024 hack
So far we have an end to end demonstration of:
- placing a battle building down on a map
- two players playing the Sky Strife match that's associated with the
  tile
- The winner of the Sky Strife match gaining possession of the tile with
  it lighting up their team's colour

Unfortunately we ran out of time to make it work for more than one
battle because the match counter is incremented on the
battle building instance coupled to a tile  which means if you place another
building down it'll try and create the same match ID as before and break.
This can be fixed by keeping the counter on an interim contract.

Main difference to this committed  version of the hack to the one we had on the day
is that there is an deploy Anvil deploy script. Making it work during the hack was a
painful experience as there was a manual process of...

- deploy the Downstream buildings
- place them on the map
- Look in the console for contract address attached to buildings
- running a separate script to get the Sky Strife Orb token address
- Transferring Orbs using metamask to the battle building
- oh and remembering to update the hard coded 'first match in window'
  value that's part of the sky strife match system

Essentially there was a lot of manual labour to make it work, no idea
why we thought it was quicker to not bother with a deploy script.